Added admin Repo and IAdminRepo because of seperation of concern
